Made RigidBody compatible with forces also built a little bit of the force registration system

+++ include/Vector2.hpp
@@ -11,9 +11,9 @@ class Vector2{
 
     Vector2 operator-(const Vector2& other) const ;
 
-    Vector2 operator*(const int scalar) const ;
+    Vector2 operator*(const float scalar) const ;
 
-    Vector2 operator/(const int scalar) const ;
+    Vector2 operator/(const float scalar) const ;
     
     Vector2& operator+=(const Vector2& other) ;
     Vector2& operator-=(const Vector2& other) ;
