+++ src/Vector2.cpp
@@ -5,10 +5,10 @@ Vector2 Vector2::operator+(const Vector2& other) const {
 Vector2 Vector2::operator-(const Vector2& other) const {
     return Vector2(x - other.x, y - other.y);
 }
-Vector2 Vector2::operator*(const int scalar) const {
+Vector2 Vector2::operator*(const float scalar) const {
     return Vector2(x * scalar, y * scalar);
 }
-Vector2 Vector2::operator/(const int scalar) const {
+Vector2 Vector2::operator/(const float scalar) const {
     return Vector2(x / scalar, y / scalar);
 }
 Vector2& Vector2::operator+=(const Vector2& other) {
